backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
PyJWT==2.9.0
Werkzeug==3.0.3
opencv-python==4.10.0.84
numpy==2.0.1
Pillow==10.4.0
reportlab==4.2.2
